WIP on testing
Added the beginning of a test. The test currently fails at the SPAdes
execution stage because SPAdes fails with a cryptic error. I get the
same error with JGI reads. Sent email to SPAdes help.

Various debugging in the impl file.

Updated dockerfile to get rid of python security warnings.